+++ packages/sit-onyx/src/components/OnyxRadioButtonGroup/types.ts
@@ -1,7 +1,7 @@
 import type { DensityProp } from "../../composables/density";
 import type { RequiredMarkerProp } from "../../composables/required";
 import type { CustomValidityProp } from "../../composables/useCustomValidity";
-import type { Direction, SelectOption, SelectOptionValue } from "../../types";
+import type { BaseSelectOption, Direction, SelectOptionValue } from "../../types";
 
 export type OnyxRadioButtonGroupProps<TValue extends SelectOptionValue = SelectOptionValue> =
   DensityProp &
@@ -41,6 +41,6 @@ export type OnyxRadioButtonGroupProps<TValue extends SelectOptionValue = SelectO
     };
 
 export type RadioButtonOption<TValue extends SelectOptionValue = SelectOptionValue> = Omit<
-  SelectOption<TValue>,
+  BaseSelectOption<TValue>,
   "hideLabel"
 >;