+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # alpha-test-app
 
+## 0.1.0-alpha.113
+
+### Patch Changes
+
+- Updated dependencies [b0554c8]
+- Updated dependencies [b0554c8]
+  - sit-onyx@1.0.0-alpha.116
+
+## 0.1.0-alpha.112
+
+### Patch Changes
+
+- Updated dependencies [19011d6]
+- Updated dependencies [19011d6]
+  - sit-onyx@1.0.0-alpha.115
+
 ## 0.1.0-alpha.111
 
 ### Patch Changes