+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,21 @@
 # playground
 
+## 0.0.1-alpha.55
+
+### Patch Changes
+
+- Updated dependencies [b0554c8]
+- Updated dependencies [b0554c8]
+  - sit-onyx@1.0.0-alpha.116
+
+## 0.0.1-alpha.54
+
+### Patch Changes
+
+- Updated dependencies [19011d6]
+- Updated dependencies [19011d6]
+  - sit-onyx@1.0.0-alpha.115
+
 ## 0.0.1-alpha.53
 
 ### Patch Changes